==> Python/2-AccuracySetup-v032824.ipynb <==
# Dependencies
import pandas as pd
import numpy as np

# Store filepath in a variable
df = pd.read_csv("Resources/cleaned_data.csv")

df.columns.tolist()

#Make a dataframe to store accuracy scores
acc_scores = (
    df[[
        'RA',
        'PID',
        'Arm',
        'Week',
        'redcap_event_name',
        'Target',
        'Word_ID',
        'Production',
        'Prod_Arpabet',
        'Prod_Word_N',
        'Prod_Phon_N',
        'Prod_Word_IPA',
        'Prod_Phon_IPA',
        'Prod_Phoneme_ID',
        'Prod_voice',
        'Prod_Place',
        'Prod_Manner',
        'Prod_tense',
        'Prod_round',
        'Prod_Height',
        'Prod_Frontness',        
        'Target_Arpabet',
        'Target_Word_IPA',
        'Target_N_Tot_Characters',
        'Target_Phon_Arpabet',
        'Target_Phoneme_ID',
        'Target_Word_NumID',
        'Target_Phon_IPA',
        'Target_voice',
        'Target_Place',
        'Target_Manner',
        'Target_tense',
        'Target_round',
        'Target_Height',
        'Target_Frontness',
        'wab1_aq',
        'wab1_ss_total',
        'wab1_avc_total',
        'wab1_r_total',
        'wab1_nwf_total',
        'wabaq_start'
        ]]
        ).copy()


# apply function to calculate voicing accuracy
acc_scores['Phon_Acc'] = df.apply(lambda x: 1 if x['Prod_Phoneme_ID'] == x['Target_Phoneme_ID'] else 0, axis=1)
acc_scores.head()

# apply function to calculate voicing accuracy
acc_scores['Voicing_Acc'] = df.apply(lambda x: 1 if x['Prod_voice'] == x['Target_voice'] else 0, axis=1)
acc_scores.head()

# apply function to calculate place accuracy
acc_scores['Place_Acc'] = df.apply(lambda x: 1 if x['Prod_Place_N'] == x['Target_Place_N'] else 0, axis=1)
acc_scores.head()

# apply function to calculate manner accuracy
acc_scores['Manner_Acc'] = df.apply(lambda x: 1 if x['Prod_Manner_N'] == x['Target_Manner_N'] else 0, axis=1)
acc_scores.head()

# apply function to calculate height accuracy
acc_scores['Height_Acc'] = df.apply(lambda x: 1 if x['Prod_Height_N'] == x['Target_Height_N'] else 0, axis=1)
acc_scores.head()

# apply function to calculate frontness accuracy
acc_scores['Frontness_Acc'] = df.apply(lambda x: 1 if x['Prod_Frontness_N'] == x['Target_Frontness_N'] else 0, axis=1)
acc_scores.head()

# apply function to calculate tenseness accuracy
acc_scores['Tenseness_Acc'] = df.apply(lambda x: 1 if x['Prod_tense'] == x['Target_tense'] else 0, axis=1)
acc_scores.head()

# apply function to calculate roundness accuracy
acc_scores['Roundness_Acc'] = df.apply(lambda x: 1 if x['Prod_round'] == x['Target_round'] else 0, axis=1)
acc_scores.head()

#Save accuracy scores
acc_scores.to_csv("Resources/acc_scores-v2.csv", index=False, header=True)

#Do the same thing for the df
df['Phon_Acc'] = df.apply(lambda x: 1 if x['Prod_Phoneme_ID'] == x['Target_Phoneme_ID'] else 0, axis=1)
df['Voicing_Acc'] = df.apply(lambda x: 1 if x['Prod_voice'] == x['Target_voice'] else 0, axis=1)
df['Place_Acc'] = df.apply(lambda x: 1 if x['Prod_Place_N'] == x['Target_Place_N'] else 0, axis=1)
df['Manner_Acc'] = df.apply(lambda x: 1 if x['Prod_Manner_N'] == x['Target_Manner_N'] else 0, axis=1)
df['Height_Acc'] = df.apply(lambda x: 1 if x['Prod_Height_N'] == x['Target_Height_N'] else 0, axis=1)
df['Frontness_Acc'] = df.apply(lambda x: 1 if x['Prod_Frontness_N'] == x['Target_Frontness_N'] else 0, axis=1)
df['Tenseness_Acc'] = df.apply(lambda x: 1 if x['Prod_tense'] == x['Target_tense'] else 0, axis=1)
df['Roundness_Acc'] = df.apply(lambda x: 1 if x['Prod_round'] == x['Target_round'] else 0, axis=1)
df.head()

#Save the aggregated dataframe
df.to_csv("Resources/baseline_data_pvm_acc.csv", index=False, header=True)
