==> Python/3-FeatureAccuracy-v032824.ipynb <==
# Dependencies
import pandas as pd
import numpy as np

# Store filepath in a variable
dfPVM = pd.read_csv("Resources/baseline_data_pvm_acc.csv")

# Save columns to list to reference later
dfPVM.columns.tolist()

df = dfPVM[[
    'Phon_Sess_Code',
    'Phon_Acc',
    'Prod_syllabic',
    'Prod_consonantal',
    'Prod_sonorant',
    'Prod_continuant',
    'Prod_delayed release',
    'Prod_approximant',
    'Prod_tap',
    'Prod_nasal',
    'Prod_voice',
    'Prod_spread gl',
    'Prod_constr gl',
    'Prod_labial',
    'Prod_round',
    'Prod_labiodental',
    'Prod_coronal',
    'Prod_anterior',
    'Prod_distributed',
    'Prod_strident',
    'Prod_lateral',
    'Prod_dorsal',
    'Prod_high',
    'Prod_low',
    'Prod_front',
    'Prod_back',
    'Prod_tense',
    'Prod_lax',
    'Prod_vowel',
    'Prod_consonant',
    'Prod_diphthong',
    'Prod_monophthong',
    'Prod_velar',
    'Prod_alveolar',
    'Prod_post-alveolar',
    'Prod_dental',
    'Prod_palatal',
    'Prod_glottal',
    'Prod_stop',
    'Prod_fricative',
    'Prod_affricate',
    'Prod_glide',
    'Prod_Place',
    'Prod_Manner',
    'Prod_Place_N',
    'Prod_Manner_N',
    'Prod_Height',
    'Prod_Frontness',
    'Prod_Height_N',
    'Prod_Frontness_N',
    'Target_syllabic',
    'Target_consonantal',
    'Target_sonorant',
    'Target_continuant',
    'Target_delayed release',
    'Target_approximant',
    'Target_tap',
    'Target_nasal',
    'Target_voice',
    'Target_spread gl',
    'Target_constr gl',
    'Target_labial',
    'Target_round',
    'Target_labiodental',
    'Target_coronal',
    'Target_anterior',
    'Target_distributed',
    'Target_strident',
    'Target_lateral',
    'Target_dorsal',
    'Target_high',
    'Target_low',
    'Target_front',
    'Target_back',
    'Target_tense',
    'Target_lax',
    'Target_vowel',
    'Target_consonant',
    'Target_diphthong',
    'Target_monophthong',
    'Target_velar',
    'Target_alveolar',
    'Target_post-alveolar',
    'Target_dental',
    'Target_palatal',
    'Target_glottal',
    'Target_stop',
    'Target_fricative',
    'Target_affricate',
    'Target_glide'
]].copy()

df.head()

# Getting a list of features from the DataFrame
features = [col.split("_", 1)[-1] for col in df.columns if col.startswith('Prod_')]

# Calculating the Conditional Difference Scores for each feature
for feature in features:
    prod_col = 'Prod_' + feature
    target_col = 'Target_' + feature

    if prod_col in df.columns and target_col in df.columns:
        acc_column_name = feature + '_Acc'
        # Set acc_column_name to 1 where prod_col and target_col are equal, and to 0 otherwise
        df.loc[:, acc_column_name] = (df[prod_col] == df[target_col]).astype(int)

# Print or save the resulting DataFrame
df[df['Phon_Acc']==0]

df.columns.to_list()

# Sum up the feature accuracy scores and divide by the total number of features the words could share (40) 

column_names = [
    'syllabic_Acc',
    'consonantal_Acc',
    'sonorant_Acc',
    'continuant_Acc',
    'delayed release_Acc',
    'approximant_Acc',
    'tap_Acc',
    'nasal_Acc',
    'voice_Acc',
    'spread gl_Acc',
    'constr gl_Acc',
    'labial_Acc',
    'round_Acc',
    'labiodental_Acc',
    'coronal_Acc',
    'anterior_Acc',
    'distributed_Acc',
    'strident_Acc',
    'lateral_Acc',
    'dorsal_Acc',
    'high_Acc',
    'low_Acc',
    'front_Acc',
    'back_Acc',
    'tense_Acc',
    'lax_Acc',
    'vowel_Acc',
    'consonant_Acc',
    'diphthong_Acc',
    'monophthong_Acc',
    'velar_Acc',
    'alveolar_Acc',
    'post-alveolar_Acc',
    'dental_Acc',
    'palatal_Acc',
    'glottal_Acc',
    'stop_Acc',
    'fricative_Acc',
    'affricate_Acc',
    'glide_Acc'
    ]
df['FeatureWeighted_PhonAcc']= (df[column_names].sum(axis=1))/40
df.head()

# Sum up the PVM accuracy scores and divide by the total number of PVM qualities the words could share (7)

column_names2 = [
    'Voicing_Acc',
    'Place_Acc',
    'Manner_Acc',
    'Height_Acc',
    'Frontness_Acc',
    'Tenseness_Acc',
    'Roundness_Acc'
 ]

df['PVMWeighted_PhonAcc']= (dfPVM[column_names2].sum(axis=1))/7
print(df['PVMWeighted_PhonAcc'].unique())

# Combine into a new dataset
df2 = df[[
    'Phon_Sess_Code',
    'syllabic_Acc',
    'consonantal_Acc',
    'sonorant_Acc',
    'continuant_Acc',
    'delayed release_Acc',
    'approximant_Acc',
    'tap_Acc',
    'nasal_Acc',
    'voice_Acc',
    'spread gl_Acc',
    'constr gl_Acc',
    'labial_Acc',
    'round_Acc',
    'labiodental_Acc',
    'coronal_Acc',
    'anterior_Acc',
    'distributed_Acc',
    'strident_Acc',
    'lateral_Acc',
    'dorsal_Acc',
    'high_Acc',
    'low_Acc',
    'front_Acc',
    'back_Acc',
    'tense_Acc',
    'lax_Acc',
    'vowel_Acc',
    'consonant_Acc',
    'diphthong_Acc',
    'monophthong_Acc',
    'velar_Acc',
    'alveolar_Acc',
    'post-alveolar_Acc',
    'dental_Acc',
    'palatal_Acc',
    'glottal_Acc',
    'stop_Acc',
    'fricative_Acc',
    'affricate_Acc',
    'glide_Acc',
    'FeatureWeighted_PhonAcc',
    'PVMWeighted_PhonAcc'
    ]].copy()

# Show results
print(f'Raw Accuracy: {df["Phon_Acc"].mean()}')
print(f'Feature Weighted Accuracy: {df2["FeatureWeighted_PhonAcc"].mean()}')
print(f'PVM Weighted Accuracy: {df2["PVMWeighted_PhonAcc"].mean()}')

df3 = dfPVM.merge(df2, how='left', on ='Phon_Sess_Code').drop_duplicates()
df3.head()

#Save it
df3.to_csv('Resources/baseline_data_pvm_acc2.csv', index=False)
