==> Python/4-EditDistances-v032824.ipynb <==
### Import dependencies

# Dependencies
import pandas as pd
import numpy as np
import textdistance as td
import panphon as pp

# Make sure you can see all output
pd.options.display.max_rows = 4000

### Data set up

# Store filepath in a variable
df = pd.read_csv("Resources/baseline_data_pvm_acc2.csv")

df.head()

# Check columns
df.columns.tolist()

### Damerau Levenshtein Edit Distance

#Check 
df.Target_Word_IPA.unique()

df[df['Target_Word_IPA'].isna()][['Prod_Word_IPA','Target_Word_IPA','Prod_Phon_IPA', 'Target_Phon_IPA']]

#Get number of phonemes for each production
df['Prod_N_Tot_Phonemes'] = (
    df
    .apply(
        lambda x:
        len(x['Prod_Word_IPA']),
        axis=1
    )
)

#Get number of phonemes for each target
df['Target_N_Tot_Characters'] = (
    df
    .apply(
        lambda x:
        len(x['Target_Word_IPA']),
        axis=1
    )
)


# Calculate the Damerau_Levenshtein 
df['Damerau_Levenshtein'] = (
    df
    .apply(
        lambda x:
        td.damerau_levenshtein(str(x['Target_Word_IPA']),str(x['Prod_Word_IPA'])),
        axis=1
        )
)

df['Damerau_Levenshtein'].mean()

print(sorted(df['Target_N_Tot_Characters'].unique()))
print(sorted(df['Damerau_Levenshtein'].unique()))

### Using PanPhon

# Check IPA symbols for fit with PanPhon
df['Target_Phon_IPA'].unique()

df[df['Target_Phon_IPA'].isna()][['Prod_Word_IPA','Target_Word_IPA','Prod_Phon_IPA', 'Target_Phon_IPA']]

df.columns.to_list()

# Create Bins based on phonological accuracy

#Create a list of subject Raw Phonological Accuracy scores at Baseline
BasePhonAcc = df[df['Session_ID']==0].groupby('PID')['Phon_Acc'].mean().to_frame().reset_index()
BasePhonAcc

# Create bins in which to place values based on Baseline Phon_Acc scores
bins = [
    0, 
    .05, 
    .1, 
    .15, 
    .2, 
    .25, 
    .3, 
    .35, 
    .4, 
    .45, 
    .5, 
    .55, 
    .6, 
    .65, 
    .7, 
    .75, 
    .8, 
    .85, 
    .9, 
    .95, 
    1
        ]

# Create labels for these bins
group_labels = [
    '1-5%', 
    '6-10%', 
    '11-15%', 
    '16-20%', 
    '21-25%',
    '26-30%',
    '31-35%', 
    '36-40%',
    '41-45%',
    '46-50%',
    '51-55%',
    '56-60%',
    '61-65%',
    '76-70%',
    '71-75%',
    '76-80%',
    '81-85%',
    '86-90%',
    '91-95%',
    '96-100%'
    ]

# Slice the data and place it into bins
BasePhonAcc['Phon_Acc_start']=pd.cut(BasePhonAcc["Phon_Acc"], bins, labels=group_labels)
BasePhonAcc=BasePhonAcc.drop(columns=['Phon_Acc'])
BasePhonAcc

#Merge in starting Phon_Acc_start score
df2 = df.merge(BasePhonAcc, on='PID', how='left')
df2['Phon_Acc_start'] = df2['Phon_Acc_start'].astype(str)
df2.head()

### Save file

df2.to_csv('Resources/baseline_data_pvm_acc3.csv', index=False)

## Just Outcome Variables

dfIVDVs = (
    df2[[
        'PID',
        'Arm',
        'Week',
        'Target',
        'Production',
        'Word_ID',
        'Session_ID',
        'Prod_Word_N',
        'Prod_Phon_N',
        'Phon_Sess_Code',
        'Word_Sess_Code',
        'Prod_Word_IPA',
        'Prod_Phon_IPA',
        'Target_Word_IPA',
        'Target_Phon_IPA',
        'Prod_voice',
        'Prod_Place',
        'Prod_Manner',
        'Prod_Height',
        'Prod_Frontness',
        'Target_voice',
        'Target_Place',
        'Target_Manner',
        'Target_Height',
        'Target_Frontness',
        'wab1_aq',
        'wabaq_start',
        'Phon_Acc',
        'Voicing_Acc',
        'Place_Acc',
        'Manner_Acc',
        'Height_Acc',
        'Frontness_Acc',
        'Tenseness_Acc',
        'Roundness_Acc',
        'syllabic_Acc',
        'consonantal_Acc',
        'sonorant_Acc',
        'continuant_Acc',
        'delayed release_Acc',
        'approximant_Acc',
        'tap_Acc',
        'nasal_Acc',
        'voice_Acc',
        'spread gl_Acc',
        'constr gl_Acc',
        'labial_Acc',
        'round_Acc',
        'labiodental_Acc',
        'coronal_Acc',
        'anterior_Acc',
        'distributed_Acc',
        'strident_Acc',
        'lateral_Acc',
        'dorsal_Acc',
        'high_Acc',
        'low_Acc',
        'front_Acc',
        'back_Acc',
        'tense_Acc',
        'lax_Acc',
        'vowel_Acc',
        'consonant_Acc',
        'diphthong_Acc',
        'monophthong_Acc',
        'velar_Acc',
        'alveolar_Acc',
        'post-alveolar_Acc',
        'dental_Acc',
        'palatal_Acc',
        'glottal_Acc',
        'stop_Acc',
        'fricative_Acc',
        'affricate_Acc',
        'glide_Acc',
        'Damerau_Levenshtein',
        'Phon_Acc_start'
    ]]
)

dfIVDVs.head()

dfIVDVs.groupby('PID')['Phon_Sess_Code'].nunique()

dfIVDVs.to_csv('Resources/lvPhonBaselineIVsDVs.csv', index=False)
