==> Python/4-EditDistances-v041224.ipynb <==
### Import dependencies

# Dependencies
import pandas as pd
import numpy as np
import textdistance as td
import panphon as pp

# Make sure you can see all output
pd.options.display.max_rows = 4000

### Data set up

# Store filepath in a variable
df = pd.read_csv("Resources/baseline_data_pvm_acc2.csv")

df.head()

df.PID.sort_values().unique()

# Check columns
df.columns.tolist()

### Damerau Levenshtein Edit Distance

#Check 
df.Target_Word_IPA.unique()

df[df['Target_Word_IPA'].isna()][['Prod_Word_IPA','Target_Word_IPA','Prod_Phon_IPA', 'Target_Phon_IPA']]

#Get number of phonemes for each production
df['Prod_N_Tot_Phonemes'] = (
    df
    .apply(
        lambda x:
        len(x['Prod_Word_IPA']),
        axis=1
    )
)

#Get number of phonemes for each target
df['Target_N_Tot_Characters'] = (
    df
    .apply(
        lambda x:
        len(x['Target_Word_IPA']),
        axis=1
    )
)


# Calculate the Damerau_Levenshtein 
df['Damerau_Levenshtein'] = (
    df
    .apply(
        lambda x:
        td.damerau_levenshtein(str(x['Target_Word_IPA']),str(x['Prod_Word_IPA'])),
        axis=1
        )
)

df['Damerau_Levenshtein'].mean()

print(sorted(df['Target_N_Tot_Characters'].unique()))
print(sorted(df['Damerau_Levenshtein'].unique()))

### Using PanPhon

# Check IPA symbols for fit with PanPhon
df['Target_Phon_IPA'].unique()

df[df['Target_Phon_IPA'].isna()][['Prod_Word_IPA','Target_Word_IPA','Prod_Phon_IPA', 'Target_Phon_IPA']]

df.columns.to_list()

# Create Bins based on phonological accuracy

# Create a data frame with mean scores for Accuracy at Baseline
BaseAcc = df[df['Session_ID'] == 0].groupby('PID').agg({
    'Phon_Acc': 'mean',  # Calculate mean Phonological Accuracy
    'Manner_Acc': 'mean',  # Calculate mean Manner of Articulation Accuracy
    'Place_Acc': 'mean', #Calulate mean Place of Articulation Accuracy
    'Voicing_Acc': 'mean', #Calulate mean Place of Articulation Accuracy
    'Damerau_Levenshtein': 'mean'
}).reset_index()

# Display the resulting data frame
BaseAcc

# Create bins in which to place values based on Baseline Phon_Acc scores
bins = [
    0, 
    .05, 
    .1, 
    .15, 
    .2, 
    .25, 
    .3, 
    .35, 
    .4, 
    .45, 
    .5, 
    .55, 
    .6, 
    .65, 
    .7, 
    .75, 
    .8, 
    .85, 
    .9, 
    .95, 
    1
        ]

# Create labels for these bins
group_labels = [
    '1-5%', 
    '6-10%', 
    '11-15%', 
    '16-20%', 
    '21-25%',
    '26-30%',
    '31-35%', 
    '36-40%',
    '41-45%',
    '46-50%',
    '51-55%',
    '56-60%',
    '61-65%',
    '66-70%',
    '71-75%',
    '76-80%',
    '81-85%',
    '86-90%',
    '91-95%',
    '96-100%'
    ]

# Slice the data and place it into bins
BaseAcc['Phon_Acc_Group']=pd.cut(BaseAcc["Phon_Acc"], bins, labels=group_labels)
BaseAcc['Place_Acc_Group']=pd.cut(BaseAcc["Place_Acc"], bins, labels=group_labels)
BaseAcc['Manner_Acc_Group']=pd.cut(BaseAcc["Manner_Acc"], bins, labels=group_labels)
BaseAcc['Voicing_Acc_Group']=pd.cut(BaseAcc["Voicing_Acc"], bins, labels=group_labels)

BaseAcc

# Save it
BaseAcc.to_csv('Resources/lvPhonBaseAcc.csv', index=False)

# Reduce the dataset to just contain the groups
BaseAcc2 = BaseAcc[['PID','Phon_Acc_Group','Place_Acc_Group','Manner_Acc_Group','Voicing_Acc_Group']]
BaseAcc2

#Merge in starting Phon_Acc_start score
df2 = df.merge(BaseAcc2, on='PID', how='left')
df2.head()

### Save file

df2.to_csv('Resources/baseline_data_pvm_acc3.csv', index=False)

# Check sessions, should only be baseline
df2.Week.unique()

# Merge MRI data with the Accuracy Data

# Load MRI data
MRIdat = pd.read_csv("Resources/lvPPA-MRI-data.csv")
MRIdat.head()

# Check out the participants
MRIdat.PID.unique()

# Read in region names
BNA = pd.read_csv("Resources/BNA_subregions.csv")
BNA.head()

# Rename ID Number columns in both dataframes
MRIdat = (
    MRIdat
    .rename(columns={
        'label':'Region_ID_Num'
        })
)

BNA = (
    BNA
    .rename(columns={
        'Label ID':'Region_ID_Num'
        })
)

# Merge anatomical information into the participant dataset
MRIdat2 = pd.merge(MRIdat, BNA, on='Region_ID_Num', how='left')
MRIdat2.head()

# Merge in the BaseAcc data
MRIdat3 = MRIdat2.merge(BaseAcc, on='PID', how='left')
MRIdat3.head()

MRIdat3.to_csv('Resources/lvPhon-MRI-Acc-summary-dat.csv', index=False)

# Different merge style, should take a while
df3 = df2.merge(MRIdat2, on='PID', how='left')
df3.head()

#check columns
df3.columns.to_list()

# Save it
df3.to_csv('Resources/lvPhon-MRI-Acc-full-dat.csv')
